Fixed bugs and ready update code
1. Restored missed inline(c++ keyword) specifier

Documentation from the official developers of the c++ programming language:
An inline function or inline variable (since C++17) has the following properties:

The definition of an inline function or variable (since C++17) must be reachable in the translation unit where it is accessed (not necessarily before the point of access).
An inline function or variable (since C++17) with external linkage (e.g. not declared static) has the following additional properties:
There may be more than one definition of an inline function or variable (since C++17) in the program as long as each definition appears in a different translation unit and (for non-static inline functions and variables (since C++17)) all definitions are identical. For example, an inline function or an inline variable (since C++17) may be defined in a header file that is #include'd in multiple source files.
It must be declared inline in every translation unit.
It has the same address in every translation unit.

2.Fixed other little problems in the source code

1. 恢复了遗漏的inline(c++关键字)说明符

文档来自c++编程语言的官方开发者。
一个内联函数或内联变量（自C++17以来）具有以下属性。

一个内联函数或变量（自C++17以来）的定义必须在它被访问的翻译单元中可以到达（不一定在访问点之前）。
一个具有外部链接的内联函数或变量（自C++17以来）（例如，未声明为静态的）具有以下附加属性。
只要每个定义出现在不同的翻译单元中，并且（对于非静态的内联函数和变量（自C++17以来））所有的定义都是相同的，那么程序中可以有不止一个内联函数或变量（自C++17以来）的定义。例如，一个内联函数或内联变量（自C++17以来）可能被定义在一个头文件中，该文件被#include'd在多个源文件中。
它必须在每个翻译单元中被内联声明。
它在每个翻译单元中都有相同的地址。

2.修复了源代码中的其他小问题